requirements.txt
py-cord
sqlalchemy
python-dotenv
asyncpg
pydantic